+++ changelog.md
@@ -33,4 +33,7 @@
 
 - added `td.partition` function
 - delouser search/replace added to configuration (undocumented)
-- `id` and `did` (document id) variables added for conversion
+- `id` and `docId` variables added for conversion
+- `config.onAction` added for parsing step notification
+- `config.onConvert` added for conversion notification
+